Localize version_env for crates deployment
